requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
arrow==1.3.0
attrs==24.2.0
cachetools==5.5.0
cartesia==1.0.14
certifi==2024.8.30
charset-normalizer==3.4.0
coloredlogs==15.0.1
daily-python==0.11.0
dataclasses-json==0.6.7
deepgram-sdk==3.7.5
deprecation==2.1.0
distro==1.9.0
flatbuffers==24.3.25
flexcache==0.3
flexparser==0.3.1
frozenlist==1.5.0
future==1.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-texttospeech==2.17.2
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
humanfriendly==10.0
idna==3.10
iniconfig==2.0.0
iterators==0.2.0
jiter==0.6.1
loguru==0.7.2
Markdown==3.7
marshmallow==3.23.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
onnxruntime==1.19.2
openai==1.50.2
packaging==24.1
pillow==10.4.0
Pint==0.24.3
pipecat==0.3.0
pipecat-ai==0.0.47
pluggy==1.5.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
pyloudnorm==0.1.1
pyparsing==3.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-deepcompare==1.0.1
python-dotenv==1.0.1
python-runner==0.6.1
requests==2.32.3
rsa==4.9
Runner==1.1
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
sympy==1.13.3
tqdm==4.66.5
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
websockets==13.1
yarl==1.16.0
Flask==3.0.3
vertexai==1.70.0
